chore(backend): satisfy ruff in dashboard module
- pyproject.toml: extend B008 per-file-ignore to dashboard/api/** (same
  convention used for src/nowgo_saude/api/**). FastAPI's Depends() and
  Query() in default-arg position is idiomatic and unavoidable.
- aggregations.py, kpis.py, units.py: drop redundant int() wrapping
  around round() — round(x) already returns int in Python 3 (RUF046).

ruff check src tests alembic now passes clean; unblocks backend-ci.yml
on this branch (previously red since PR #1).

+++ backend/src/nowgo_saude/dashboard/services/aggregations.py
@@ -102,7 +102,7 @@ def heatmap_by_ra(session: Session) -> list[RegionPressureOut]:
         score_raw = density * 0.6 + (sev_sum / max(len(recent), 1)) * 18 + (
             neg_count / max(len(recent), 1)
         ) * 25
-        pressure = max(0, min(100, int(round(score_raw))))
+        pressure = max(0, min(100, round(score_raw)))
         delta = len(recent) - len(prior)
         trend = "up" if delta > max(3, len(prior) * 0.15) else (
             "down" if delta < -max(3, len(prior) * 0.15) else "stable"
@@ -203,7 +203,7 @@ def region_detail(session: Session, ra_id: str) -> RegionDetailOut | None:
     score_raw = density * 0.6 + (sev_sum / max(len(recent), 1)) * 18 + (
         neg_count / max(len(recent), 1)
     ) * 25
-    pressure = max(0, min(100, int(round(score_raw))))
+    pressure = max(0, min(100, round(score_raw)))
     delta = len(recent) - len(prior)
     trend = "up" if delta > max(3, len(prior) * 0.15) else (
         "down" if delta < -max(3, len(prior) * 0.15) else "stable"
@@ -312,7 +312,7 @@ def unit_detail(session: Session, unit_id: str) -> UnitDetailOut | None:
         + min(max(growth_pct, 0), 120) * 0.25
         + (neg / max(len(recent), 1)) * 18
     )
-    score = max(0, min(100, int(round(raw))))
+    score = max(0, min(100, round(raw)))
 
     topic_counts: dict[str, int] = defaultdict(int)
     for _, t, _, _, _, _ in recent:
